Update __init__.py
Due to more experience in the composition of the data and the changed method of interpreting data, APSystemsInvalidData only outputs a "Timeout" message. 
As a result, the importance of this logging rule has lapsed.
For replacement insight, the use of displaying the "binairy_sensor.ecu_using_cached_data" with optional secondary information "Last Changed" is suggested.