Change today_energy sensor stateclass back to TOTAL_INCREASING
According to the HA docs and the errors seen in issue #174:

Similar to total, with the restriction that the state represents a
monotonically increasing positive total which periodically restarts
counting from 0, e.g. a daily amount of consumed gas, weekly water
consumption or lifetime energy consumption.

Seems to be the correct stateclass to use.